Create func package and ComposedFunc
